+++ R/as_factor.R
@@ -80,7 +80,7 @@ as_factor <- function(x, ..., add.non.labelled = FALSE) {
 
 #' @export
 as_factor.default <- function(x, ..., add.non.labelled = FALSE) {
-  .dat <- get_dot_data(x, dplyr::quos(...))
+  .dat <- get_dot_data(x, rlang::quos(...))
   to_fac_helper(.dat, add.non.labelled)
 }
 
@@ -88,7 +88,7 @@ as_factor.default <- function(x, ..., add.non.labelled = FALSE) {
 #' @export
 as_factor.data.frame <- function(x, ..., add.non.labelled = FALSE) {
   # evaluate arguments, generate data
-  .dat <- get_dot_data(x, dplyr::quos(...))
+  .dat <- get_dot_data(x, rlang::quos(...))
 
   for (i in colnames(.dat)) {
     x[[i]] <- to_fac_helper(.dat[[i]], add.non.labelled)