+++ R/drop_labels.R
@@ -2,7 +2,7 @@
 #' @export
 drop_labels <- function(x, ..., drop.na = TRUE) {
   # evaluate arguments, generate data
-  .dat <- get_dot_data(x, dplyr::quos(...))
+  .dat <- get_dot_data(x, rlang::quos(...))
 
   if (is.data.frame(x)) {
     # iterate variables of data frame