+++ R/set_na.R
@@ -143,7 +143,7 @@ set_na <- function(x, ..., na, drop.levels = TRUE, as.tag = FALSE) {
   }
 
   # evaluate arguments, generate data
-  .dat <- get_dot_data(x, dplyr::quos(...))
+  .dat <- get_dot_data(x, rlang::quos(...))
 
   if (is.data.frame(x)) {
     # iterate variables of data frame