+++ R/tidy_labels.R
@@ -38,7 +38,7 @@
 #' @export
 tidy_labels <- function(x, ..., sep = "_", remove = FALSE) {
   # evaluate arguments, generate data
-  .dat <- get_dot_data(x, dplyr::quos(...))
+  .dat <- get_dot_data(x, rlang::quos(...))
 
   if (is.data.frame(x)) {
     # iterate variables of data frame