+++ R/zap_labels.R
@@ -117,7 +117,7 @@
 #' @export
 zap_labels <- function(x, ...) {
   # evaluate arguments, generate data
-  .dat <- get_dot_data(x, dplyr::quos(...))
+  .dat <- get_dot_data(x, rlang::quos(...))
 
   if (is.data.frame(x)) {
     # iterate variables of data frame
@@ -137,7 +137,7 @@ zap_labels <- function(x, ...) {
 #' @export
 zap_unlabelled <- function(x, ...) {
   # evaluate arguments, generate data
-  .dat <- get_dot_data(x, dplyr::quos(...))
+  .dat <- get_dot_data(x, rlang::quos(...))
 
   if (is.data.frame(x)) {
     # iterate variables of data frame
@@ -187,7 +187,7 @@ zap_unlabelled <- function(x, ...) {
 #' @export
 zap_na_tags <- function(x, ...) {
   # evaluate arguments, generate data
-  .dat <- get_dot_data(x, dplyr::quos(...))
+  .dat <- get_dot_data(x, rlang::quos(...))
 
   if (is.data.frame(x)) {
     # iterate variables of data frame